Version 3 - Add login and register controllers that validate JWT - Implement User mongoose model - Add pre-save to the users model for encrpypting a password with bcryptjs - Add methods to the users model to create JWT token and compare passwords when logging in - Create authenticatino middleware for accessing the other routes - Implement auth controllers - Modify projects controllers to get data only for the current user - Implement error handling for validation, cast and duplicate errors - Add helmet, xss and rate-limiter for extra protection

+++ server/controllers/projects.js
@@ -1,40 +1,53 @@
-import { BadRequestError, NotFoundError } from '../errors/index.js'
+import { NotFoundError } from '../errors/index.js'
 import Project from '../models/Project.js'
 
 export const getAllProjects = async (req, res) => {
-  const projects = await Project.find({})
+  const projects = await Project.find({ userId: req.user.userId }).sort(
+    'createdAt'
+  )
   res.status(200).json({ projects, count: projects.length })
 }
 
 export const createProject = async (req, res) => {
+  req.body.userId = req.user.userId
   const project = await Project.create(req.body)
   res.status(201).json({ project })
 }
 
 export const getProject = async (req, res) => {
-  const projectId = req.params.id
-  const project = await Project.findOne({ _id: projectId })
+  const {
+    user: { userId },
+    params: { id: projectId },
+  } = req
+  const project = await Project.findOne({ _id: projectId, userId })
   if (!project) throw new NotFoundError(`No job with id: ${projectId}`)
   res.status(200).json({ project })
 }
 
 export const updateProject = async (req, res) => {
   const {
-    body: { title },
+    user: { userId },
     params: { id: projectId },
   } = req
-  if (title === '') throw new BadRequestError('Title field cannot be empty')
-  const project = await Project.findOneAndUpdate({ _id: projectId }, req.body, {
-    new: true,
-    runValidators: true,
-  })
+
+  const project = await Project.findOneAndUpdate(
+    { _id: projectId, userId },
+    req.body,
+    {
+      new: true,
+      runValidators: true,
+    }
+  )
   if (!project) throw new NotFoundError(`No job with id: ${projectId}`)
   res.status(200).json({ project })
 }
 
 export const deleteProject = async (req, res) => {
-  const projectId = req.params.id
-  const project = await Project.findOneAndDelete({ _id: projectId })
+  const {
+    user: { userId },
+    params: { id: projectId },
+  } = req
+  const project = await Project.findOneAndDelete({ _id: projectId, userId })
   if (!project) throw new NotFoundError(`No job with id: ${projectId}`)
   res.status(200).send()
 }
